+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # [0.2.0] - 2023-04-07
 
 ## Added
-- shorthand comannds to jump to and set portals.
+- adds list command to see all set portals (in json format as well).
+- adds shorthand commands to jump to and set portals.
 - adds ez-init command to configure shorthand for your shell.
 
 # [0.1.1] - 2023-02-01